fixed it so that it works, removed mongoose for now, but it is ready to go now :)
