+++ controllers/userControllers.js
@@ -25,5 +25,6 @@ exports.register = function (req, res) {
 }
 
 exports.home = function (req, res) {
-  res.render('home-guest')
+  res.json({"success":true})
+  //res.render('home-guest')
 }
