+++ package.json
@@ -14,7 +14,7 @@
     "validator": "^12.1.0"
   },
   "devDependencies": {
-    "concurrently": "^4.1.0",
+    "concurrently": "^4.1.2",
     "nodemon": "^1.18.7"
   },
   "scripts": {
