+++ server.js
@@ -1,20 +1,14 @@
-const express = require('express')
-const app = express()
 
-const router = require('./router')
-const mongoose = require("mongoose");
 const dotenv = require('dotenv')
 dotenv.config()
 const mongodb = require('mongodb')
-app.use(express.urlencoded({ extended: false }))
-app.use(express.json())
 
-app.use(express.static('public'))
 
-app.use(router)
+mongodb.connect(process.env.CONNECTIONSTRING, {useNewUrlParser: true, useUnifiedTopology: true}, function(err, client) {
+  module.exports = client.db()
+  const app = require('./app')
+  app.listen(process.env.PORT, test=> console.log("listening!"))
+  
+})
 
-app.post('/login', req => console.log(req))
 
-mongoose.connect(process.env.MONGODB_URI || process.env.CONNECTIONSTRING)
-
-app.listen(process.env.PORT)
